Move to inventory-based provisioning of AWS resources.
This change introduces usage of Ansible inventory to
make provisioning cleaner. The idea is to make the
playbook run faster and use more of the orchestration
capabilities of Ansible, even when provisioning resources
in the cloud.

To further simplify things, the idea of having an environment
name has been removed. The goal of this project is to have
users testing out ansible tower be able to install an instance
in their own aws account.